requirements.txt
arrow==1.2.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
canvasapi==3.0.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
cmake==3.24.1
Django==4.1.1
django-cors-headers==3.10.0
django-environ==0.8.1
django-filter==22.1
django-storages==1.13.1
djangorestframework==3.14.0
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.4
importlib-metadata==5.0.0
imutils==0.5.4
Markdown==3.4.1
mysqlclient==2.1.0
numpy==1.23.2
opencv-python==4.6.0.66
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
zipp==3.8.1
